+++ etna/datasets/utils.py
@@ -91,13 +91,9 @@ def duplicate_data(df: pd.DataFrame, segments: Sequence[str], format: str = Data
         raise ValueError("There should be 'timestamp' column")
 
     # construct long version
-    segments_results = []
-    for segment in segments:
-        df_segment = df.copy()
-        df_segment["segment"] = segment
-        segments_results.append(df_segment)
-
-    df_long = pd.concat(segments_results, ignore_index=True)
+    n_segments, n_timestamps = len(segments), df.shape[0]
+    df_long = df.iloc[np.tile(np.arange(n_timestamps), n_segments)]
+    df_long["segment"] = np.repeat(a=segments, repeats=n_timestamps)
 
     # construct wide version if necessary
     if format_enum == DataFrameFormat.wide: